failed attempt at using xulrunner-sdk
This was an almost comical waste of time. In an effort to get stupid array support working in XPCOM, I spent most of a day upgrading the version of the mozilla sdk we're using, went from gecko-sdk (1.8) to xulrunner-sdk (1.9), finally got it to build after hunting down a compiler directive that was preventing linking, I got failures in previously working code and seemingly random crashed for my troubles. Going to roll back to gecko-sdk, just want to preserve this pain-in-the-ass for posterity or something.